\name{DS705data-package}
\alias{DS705data-package}
\alias{DS705data}
\docType{package}
\title{
\packageTitle{DS705data}
}
\description{
\packageDescription{DS705data}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DS705data}
\packageIndices{DS705data}
Type data(package="DS705data") to see what datasets are available.  To use a data set:
data(WormSheep)
WormSheep
}
\author{
\packageAuthor{DS705data}

Maintainer: \packageMaintainer{DS705data}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
data(Wormsheep)
}
